requirements.txt
psycopg2-binary
flask
minio
pymongo
Jinja2
python-dotenv